requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
datasets>=2.19.0
pillow>=10.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
umap-learn>=0.5.6
matplotlib>=3.8.0
tqdm>=4.66.0